Update drawings-gods_data-lists.ttl
For both onthology and lists:
1. Corrected order of individual questionary. 2. Corrected temples, religions of Japan (collided with the rest). 3. Corrected plenty of typos. 4. Changed Labels of some objects Passage-Visit, Campagne-Reserach Wave, etc. (Marion, we might need to do it also in the owls-part, but I did not touch it.) Christelle will send complete list.
Marion, please use the numbers 1-9 for the label and other admin properties that I do not see directly for the individual questionary. There are 122 things to place by order so far, so I list them here:
1-9 - for admin of Marion, 
10- rel identity, 
11- 45 familiarity with temples, 
46-53 japan mixed (jinja-otera)
54-58 jinjia, i.e. shinto, 
59-61 objects at home japan, 
62-77 prayer,
78 -86 could you draw?...
87-96 location, age, sex of god, meaning of the word
97 -122 admin swiss

This can be merged already.

Remains from my side: new structure of keywords, to come next Tuesday.