Resolved St- Gingolphe - St. Prex.
Dear Marion,

So here is the final version of geo:
St. Gingolphe is reattached as a 2nd collect place to the village St. Prex. (Complicated case: the kids of Vaud went for vacation in Valais, but the team wants them to be attached to St. Prex, Vaud, in any case.)
So I added "Camp Saint Gingolph" to St. Prex and deleted the old structures throughout.
Plus minor typos corrected.